app/requirements.txt
attrs==20.1.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
fastapi==0.61.0
Flask==1.1.2
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.4
h11==0.9.0
httptools==0.1.1
idna==2.10
importlib-metadata==1.7.0
iniconfig==1.0.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
pycodestyle==2.6.0
pydantic==1.6.1
pyparsing==2.4.7
pytest==6.0.1
python-dotenv==0.14.0
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.18
starlette==0.13.6
toml==0.10.1
urllib3==1.25.10
uvicorn==0.11.8
uvloop==0.14.0
websockets==8.1
Werkzeug==1.0.1
zipp==3.1.0
